# openapi/schedule-estimate.yaml
openapi: 3.0.3
info:
  title: Buildertrend Mock — Schedule, Estimate & Daily Logs API
  version: 2.0.0
  description: |
    Local-only, unauthenticated REST API backing the Schedule, Estimate and
    Daily Logs features of a Buildertrend front-end mock (see the repo's
    CAPTURE_LOG.md for what in this app is captured from the real product
    vs. invented for demo purposes). No `security` scheme is defined
    anywhere in this document — that's intentional, not an omission; this
    server only ever runs on localhost for local development/agent use.

    ERRORS: every 4xx/5xx response body is an ErrorResponse whose `error`
    string is written to be actionable — an agent framework hands that text
    back to the model, so it names the offending field and, for jobId, the
    valid values. A malformed JSON body is a 400. Any 500 means a genuine
    server bug, not a caller mistake; individual operations therefore do not
    enumerate '500'.

    UPDATES are partial: PUT (and PATCH, an undocumented alias with identical
    behaviour) change only the fields you send and leave everything else at
    its stored value. This changed in 2.0.0 — it used to be a full replace,
    which silently reset omitted fields.

    This spec is written to be consumed by an agent framework's
    OpenAPI-to-tool machinery (e.g. Google ADK's `OpenAPIToolset`), which
    turns every operation below into a callable function: `operationId`
...
    them by trial and error.
servers:
...
paths:
  /api/schedule:
    get:
      operationId: listScheduleItems
      tags: [Schedule]
      summary: List all schedule items for a job.
      description: |
        Returns every schedule item belonging to the given job, ordered by
        start date then id. Returns an empty array (not an error) if the
        job has no items yet.
      parameters:
        - $ref: '#/components/parameters/JobIdQuery'
      responses:
        '200':
          description: Array of schedule items for the job (possibly empty).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScheduleItem'
        '400':
          $ref: '#/components/responses/MissingJobId'
    post:
      operationId: createScheduleItem
      tags: [Schedule]
      summary: Create a new schedule item (task or milestone) for a job.
      description: |
        Creates a schedule item under the given job and returns it with
        server-assigned `id`, `createdBy`, `createdAt`, and `updatedAt`.

        Required fields: `jobId`, `title`, `start`. All other fields are
        optional and fall back to sensible defaults (see
        ScheduleItemInput's per-field descriptions) — this mirrors the
        real "New Schedule Item" form, where only Title/Start Date/End
        Date are marked required.

        `end` is derived when omitted: `workDays` counts WORKING days
        (Mon–Fri), so `{start: "2026-08-06", workDays: 3}` ends
        2026-08-10, skipping the weekend. Send `end` instead and
        `workDays` is recounted over that span.

        `createdBy` is always server-set to a fixed demo user regardless
        of what you send — there is no real user/auth concept in this
        mock, so don't bother supplying it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleItemInput'
            example:
              jobId: j1
              title: Final Walkthrough
              start: '2026-09-01'
              end: '2026-09-01'
      responses:
        '201':
          description: The created schedule item.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleItem'
        '400':
          description: One or more required fields (jobId, title, start, end) are missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/schedule/cascade-preview:
    post:
      operationId: previewScheduleCascade
      tags: [Schedule]
      summary: Ask what would happen if schedule dates moved — writes nothing.
      description: |
        Answers "if these items moved, what else moves, and does the job finish
        later?" **This is read-only** — no schedule item is modified. Call it
        freely; it is the intended way to reason about a date change before
        proposing one.

        USE THIS INSTEAD OF DOING THE ARITHMETIC YOURSELF. Dependencies,
        durations, weekends and the critical path interact in ways that are easy
        to get wrong by hand, and a confidently wrong date is worse than no
        answer. Anything you compute yourself will disagree with what the Gantt
        shows the user.

        Each entry in `changes` names one item and how it moves — send
        `shiftDays` to push it (WORK days, so 2 on a Thursday lands Monday),
        `start` to place it on a date, and/or `workDays` to change its duration.
        Successors then move by the same rule, preserving whatever gap they
        already had, and never being pulled earlier than they already sit.

        Send `changes: []` for a pure read: you still get `projectEnd` and
        `criticalIds` for the job as it stands.

        Add `includeAnalysis: true` for per-item float. Float is what
        distinguishes "this slips 3 days and the job finishes 3 days later" from
        "this slips 3 days into slack it already had and the finish date does not
        move" — worth asking for whenever you intend to describe the impact of a
        delay, and worth omitting otherwise since it roughly doubles the reply.

        Items marked complete are never moved. If the ripple wanted to move one,
        it is reported in `conflicts` rather than silently ignored, and the items
        behind it are reported as blocked — so say so rather than claiming the
        dates changed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CascadePreviewInput'
            example:
              jobId: j1
              changes:
                - itemId: s2
                  shiftDays: 2
      responses:
        '200':
          description: What would change. Nothing has been written.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CascadePreviewResult'
        '400':
          description: |
            Unknown jobId, an unknown itemId, or a malformed `changes` entry —
            the message names which.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: |
            This job's stored dependencies contain a loop, so no ordering exists
            and nothing can be computed from them. `cycleIds` names the loop.
            Report it — it is a data problem for a human to fix, not something to
            retry or work around.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/schedule/batch:
    post:
      operationId: applyScheduleCascade
      tags: [Schedule]
      summary: Move schedule dates for real, cascading and undoable, in one write.
      description: |
        Applies date changes and the knock-on moves they cause, as ONE atomic
        write, recording a change set so the whole thing can be undone together.

        USE THIS FOR ANY DATE CHANGE, not updateScheduleItem. Moving a bar
        almost always pushes the work behind it, and updateScheduleItem moves
        exactly one item: use it for dates and you leave the schedule
        contradicting its own dependencies — a successor starting before the
        predecessor it waits on. Doing it as several updateScheduleItem calls is
        no better; one can fail after the others have landed, and there is no
        record tying them together to undo.

        Takes the same `changes` shape as previewScheduleCascade, so the normal
        sequence is: preview it, tell the user what moves, apply the identical
        payload once they agree. Preview first for anything beyond a single
        unblocked item — the ripple is often much larger than it looks.

        `reason` is shown to the user in the schedule's history, so write it for
        them ("Pushed framing 2 days for the rain delay"), not for the log.

        The response echoes the same shape the preview returned, plus the
        `changeSet` that was recorded and the job's full refreshed item list.
        Report what actually happened from `counts` and `projectEnd` rather than
        from what you intended — completed items are never moved, and any the
        ripple wanted to move come back in `conflicts`.

        A request that turns out to change nothing is a success with
        `changeSet: null`, not an error. Nothing was written, so say so.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CascadeApplyInput'
            example:
              jobId: j1
              changes:
                - itemId: s2
                  shiftDays: 2
              reason: Pushed framing 2 days for the rain delay
      responses:
        '200':
          description: |
            Applied. `changeSet` is null when the request changed nothing.
          content:
            application/json:
              schema:
                type: object
                properties:
                  changeSet:
                    $ref: '#/components/schemas/ScheduleChangeSet'
                  items:
                    type: array
                    description: The job's full schedule after the write.
                    items:
                      $ref: '#/components/schemas/ScheduleItem'
                  plan:
                    $ref: '#/components/schemas/CascadePreviewResult'
        '400':
          description: Unknown jobId, unknown itemId, or a malformed change entry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: |
            The job's dependencies contain a loop. Nothing was written. Report it
            rather than retrying.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/workday-exceptions:
    get:
      operationId: listWorkdayExceptions
      tags: [Schedule]
      summary: Days this job does or doesn't work, beyond its standard work week.
      description: |
        Holidays and closures that BLOCK a normally-working day
        (`type: "non_workday"`), and extra days that OPEN a normally-off one
        (`type: "extra_workday"`) — a Saturday push, for instance.

        Read this before explaining a date. All schedule arithmetic already
        routes around these, so a task can appear to skip a Monday or run
        through a Saturday for a reason that is invisible in the item itself.
        If someone asks why a duration doesn't line up with the calendar, this
        is usually the answer.

        Includes both this job's own exceptions and company-wide ones (those
        come back with `jobId: ""`).

        An exception with `sameEveryYear: true` repeats annually — only its
        month and day matter, so a single row covers every Christmas and its
        stored year may be in the past.
      parameters:
        - $ref: '#/components/parameters/JobIdQuery'
      responses:
        '200':
          description: Exceptions affecting this job, earliest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkdayException'
        '400':
          $ref: '#/components/responses/MissingJobId'
...
components:
  parameters:
    JobIdQuery:
      name: jobId
      in: query
      required: true
      description: |
        The job to scope this request to, e.g. "j1". Jobs themselves are
        static frontend fixtures (src/data/jobs.js), not part of this
        API — this parameter is a soft reference with no server-side
        existence check (an unknown jobId returns an empty result, not
        a 404).
      schema:
        type: string
        example: j1
...